separate product review screen to several files and components
